% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlob.R
\name{mlob}
\alias{mlob}
\title{Multi-Level Optimal Bayes Function (MLOB)}
\usage{
mlob(
  formula,
  data,
  group,
  balancing = FALSE,
  alpha = 0.05,
  jackknife = FALSE,
  ...
)
}
\arguments{
\item{formula}{A formula specifying the model (e.g., \code{Y ~ X + C...}) where Y is the dependent variable, X is the target variable (always), and C... represents control variables.}

\item{data}{A data frame containing the variables in the formula.}

\item{group}{A positive numeric value representing the number of groups (J) in the data. The number of rows in \code{data} must be divisible by \code{group}.}

\item{balancing}{Logical. If \code{TRUE}, the data will be balanced across groups. Defaults to \code{FALSE}.}

\item{alpha}{A numeric value representing the confidence level used to calculate confidence intervals for the estimators. Defaults to \code{0.05}.}

\item{jackknife}{Logical. If \code{TRUE}, the jackknife resampling method for calculating of the standard error of between-group parameter beta_b will be applied. Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list containing the results of the Bayesian estimation including model formula, response and predictor variables, and other relevant details.
}
\description{
This function estimates between-group parameter beta_b of multilevel latent variable model with control variables
using a multivariate regularized Bayesian approach.
It is primarily designed for cases where data can be divided into multiple groups
and ensures that the data is balanced across groups.
}
\details{
This function checks whether the data is balanced (i.e., whether the same number of individuals are present
in each group). If the data is unbalanced and `balancing = FALSE`, the function will stop 
and throw an error message. 

MultiLevelOptimalBayes (MLOB) is designed for small-sample two-level latent variable models, commonly used in
psychology, education, and other disciplines that deal with hierarchical data structures.
}
\examples{
# Example usage with the iris dataset
result <- mlob(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width, data = iris, group = 3)

# View summary statistics (similar to summary of a linear model)
summary(result)
## 
## Call:
## mlob(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width, data = data, group = 3) 
## 
#' ## Summary of Coefficients:
##                     Estimate Std. Error Lower CI (95\%) Upper CI (95\%)    Z value  Pr(>|z|) Significance
## beta_b              0.5018467  0.8777217     -1.2184563       2.222150  0.5717605 0.5674842
## gamma_Petal.Length  0.7117494  0.6214602     -0.5062901       1.929789  1.1452857 0.2520908
## gamma_Petal.Width  -0.5576712  1.6291191     -3.7506860       2.635344 -0.3423145 0.7321142
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

}
